pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bigr-discovery"
version = "0.1.0"
description = "BİGR compliant network asset discovery and classification agent"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Okan Yucel"}
]

dependencies = [
    "scapy>=2.5.0",
    "zeroconf>=0.131.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "dnslib>=0.9.23",
    "dnspython>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
bigr = "bigr.cli:app"

[tool.setuptools.packages.find]
include = ["bigr*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
